+++ src/main/java/racingcar/RacingCar.java
@@ -1,5 +1,7 @@
 package racingcar;
 
+import racingcar.controller.RacingController;
+
 public class RacingCar {
 
     public static void main(String[] args) {
