+++ src/main/java/racingcar/utils/Console.java
@@ -1,4 +1,4 @@
-package racingcar;
+package racingcar.utils;
 
 import java.lang.reflect.Field;
 import java.util.Objects;
