+++ src/main/java/racingcar/view/OutputView.java
@@ -1,6 +1,6 @@
-package racingcar;
+package racingcar.view;
 
-public class Output {
+public class OutputView {
 
     private static String INPUT_CAR_NAMES_MESSAGE = "경주할 자동차 이름을 입력하세요(이름은 쉼표(,)를 기준으로 구분).";
     private static String INPUT_NUMBER_MESSAGE = "시도할 회수는 몇회인가요?";
