+++ src/test/java/racingcar/utils/ValidatorTest.java
@@ -1,9 +1,9 @@
-package racingcar;
+package racingcar.utils;
 
 import org.junit.jupiter.api.Test;
 
 import static org.assertj.core.api.Assertions.*;
-import static racingcar.Validator.*;
+import static racingcar.utils.Validator.*;
 
 public class ValidatorTest {
